MAINT rename educe.stac.{document_plus to fusion}
Note that RST-DT would need to be similarly updated
